% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g1k15.R
\docType{data}
\name{g1k15}
\alias{g1k15}
\title{Data frame with coordinates of swiss cantons (2015).}
\format{A data frame with 9216 rows and 20 variables:
\describe{
  \item{\code{long}}{x coordinate}
  \item{\code{lat}}{y coordinate}
  \item{\code{order}}{order coordinate (not used)}
  \item{\code{hole}}{logical, if the polygon is a hole or not}
  \item{\code{piece}}{piece coordinate (not used)}
  \item{\code{group}}{group coordinate}
  \item{\code{id}}{id of the polygon}
 \item{\code{KTNR}}{a numeric vector}
\item{\code{GRNR}}{a numeric vector}
\item{\code{AREA_HA}}{a numeric vector}
\item{\code{X_MIN}}{a numeric vector}
\item{\code{X_MAX}}{a numeric vector}
\item{\code{Y_MIN}}{a numeric vector}
\item{\code{Y_MAX}}{a numeric vector}
\item{\code{X_CNTR}}{a numeric vector}
\item{\code{Y_CNTR}}{a numeric vector}
\item{\code{Z_MIN}}{a numeric vector}
\item{\code{Z_MAX}}{a numeric vector}
\item{\code{Z_AVG}}{a numeric vector}
\item{\code{Z_MED}}{a numeric vector}
}}
\usage{
g1k15
}
\description{
This data can be used to plot a map of swiss cantons borders for example with ggplot2.
}
\keyword{datasets}

