% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2spatial.R
\name{tidy2sp}
\alias{tidy2sp}
\title{Transforms a data frame with map coordinates in a SpatialPolygonsDataFrame.}
\usage{
tidy2sp(.data)
}
\arguments{
\item{.data}{data frame with map coordinates of polygons}
}
\value{
an object of class SpatialPolygonsDataFrame (from package sp)
}
\description{
Starting from a data frame with map coordinates, transforms
it in an object of class SpatialPolygonsDataFrame.
}
\examples{
data(g1k15)
g1k15_sp <- tidy2sp(g1k15)
class(g1k15_sp)
str(g1k15_sp)
}

