% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2spatial.R
\name{union_polygons}
\alias{union_polygons}
\title{Aggregate Polygons in a SpatialPolygonsDataFrame object}
\usage{
union_polygons(.sp, new_id)
}
\arguments{
\item{.sp}{SpatialPolygonsDataFrame}

\item{new_id}{A vector defining the output Polygons objects, equal in length to the length of the polygons slot of .sp; it may be character, integer, or factor}
}
\value{
an object of class SpatialPolygonsDataFrame
}
\description{
Aggregate Polygons in a SpatialPolygonsDataFrame object, based on a vector (new_id) of the same length of the number of polygons in the starting SpatialPolygonsDataFrame (argument .sp).
}
\examples{
data(g1k15)

# Transform data frame g1k15 in a SpatialPolygonsDataFrame
g1k15_sp <- tidy2sp(g1k15)

# Create a new SpatialPolygonsDataFrame, aggregating the polygons based on the GRNR (grand regions)
g1k15_sp_union <- union_polygons(g1k15_sp, new_id = g1k15_sp@data$GRNR)
str(g1k15_sp_union)

}

