% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2spatial.R
\name{union_tidy}
\alias{union_tidy}
\title{Aggregate polygons in a tidy data frame with map coordinates}
\usage{
union_tidy(.data, new_id)
}
\arguments{
\item{.data}{data frame with map coordinates of polygons}

\item{new_id}{A vector defining the output Polygons objects, equal in length to the length of the polygons slot of .sp; it may be character, integer, or factor}
}
\value{
a data frame with map coordinates of aggregated polygons
}
\description{
Combine tidy2sp, union_polygons and broom::tidy, returning a tidy data frame with aggregated polygons, based on a new_id vector.
}
\examples{
data(g1k15)

# Aggregate cantons into grand regions
grand_regions <- unique(g1k15[ , c("KTNR", "GRNR")])[["GRNR"]]
g1k15_union <- union_tidy(g1k15, new_id = grand_regions)
str(g1k15_union)

}

